fcd_transendence/transcendenceapi/requirements.txt
asgiref==3.8.1
attrs==23.2.0
Automat==22.10.0
certifi==2024.7.4
cffi==1.16.0
channels==4.1.0
charset-normalizer==3.3.2
ci-info==0.3.0
click==8.1.7
configobj==5.0.8
configparser==7.0.0
constantly==23.10.4
cryptography==42.0.8
daphne==3.0.2
Django==4.2.1
django-allauth==0.63.3
django-cors-headers==4.4.0
django-extensions==3.2.3
djangorestframework==3.15.2
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.1
etelemetry==0.3.1
filelock==3.15.4
fitz==0.0.1.dev2
h2==4.1.0
hpack==4.0.0
httplib2==0.22.0
hyperframe==6.0.1
hyperlink==21.0.0
idna==3.7
incremental==22.10.0
install==1.3.5
isodate==0.6.1
looseversion==1.3.0
lxml==5.2.2
networkx==3.2.1
nibabel==5.2.1
nipype==1.8.6
numpy==2.0.0
oauthlib==3.2.2
packaging==24.1
pandas==2.2.2
pathlib==1.0.1
pillow==10.4.0
postgres==4.0
priority==1.3.0
prov==2.0.1
psycopg2-binary==2.9.9
psycopg2-pool==1.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydot==2.0.0
PyJWT==1.7.1
pyOpenSSL==24.1.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
pyxnat==1.6.2
rdflib==6.3.2
requests==2.32.3
requests-oauthlib==2.0.0
scipy==1.13.1
service-identity==24.1.0
setuptools==70.3.0
simplejson==3.19.2
six==1.16.0
sqlparse==0.5.0
traits==6.3.2
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
whitenoise==6.0.0
zope.interface==6.4.post2
